added gps functionality for status

+++ app/src/main/java/com/bijgepast/locationhunter/enums/DistanceStatus.kt
@@ -1,11 +1,11 @@
 package com.bijgepast.locationhunter.enums
 
-enum class DistanceStatus(val DistanceInMeters: UInt, val nameForEnum: String) {
-    FROZEN(UInt.MAX_VALUE, "Frozen"),
-    VERY_COLD(20000u, "Very cold"),
-    COLD(15000u, "Cold"),
-    ROOM_TEMPERATURE(10000u, "Room temperature"),
-    HOT(5000u, "Hot"),
-    VERY_HOT(1000u, "Very hot"),
-    HOTTER_THEN_HELL(500u, "HOTTER THAN HELL")
+enum class DistanceStatus(val DistanceInMeters: Float, val nameForEnum: String) {
+    FROZEN(Float.MAX_VALUE, "Frozen"),
+    VERY_COLD(20000f, "Very cold"),
+    COLD(15000f, "Cold"),
+    ROOM_TEMPERATURE(10000f, "Room temperature"),
+    HOT(5000f, "Hot"),
+    VERY_HOT(1000f, "Very hot"),
+    HOTTER_THEN_HELL(500f, "HOTTER THAN HELL")
 }
